+++ R/att_from_rmds.R
@@ -9,6 +9,8 @@
 #' \dontrun{
 #' att_from_vignette("vignettes/my-vignette.Rmd")
 #' }
+#'
+#' @export
 att_from_rmd <- function(path, temp_dir = tempdir()) {
   file <- knitr::purl(path, output = paste0(tempfile(tmpdir = temp_dir), ".R"),
                       documentation = 0, quiet = TRUE)
@@ -17,7 +19,7 @@ att_from_rmd <- function(path, temp_dir = tempdir()) {
 
 #' Get all packages called in vignettes folder
 #'
-#' @param path path to directory with Rmds
+#' @param path path to directory with Rmds or vector of Rmd files
 #' @param recursive logical. Should the listing recurse into directories?
 #'
 #' @return Character vector of packages called with library or require.
@@ -30,13 +32,21 @@ att_from_rmd <- function(path, temp_dir = tempdir()) {
 #' }
 #' @export
 att_from_rmds <- function(path = "vignettes",recursive = TRUE) {
-  all_f <- list.files(path, full.names = TRUE, pattern = "*.Rmd$|*.rmd$",recursive = recursive)
-  res <- lapply(all_f, att_from_rmd) %>%
+
+  if (isTRUE(all(dir.exists(path)))) {
+    all_f <- list.files(path, full.names = TRUE, pattern = "*.Rmd$|*.rmd$",recursive = recursive)
+  } else if (isTRUE(all(file.exists(path)))) {
+    all_f <- normalizePath(path[grepl("*.Rmd$|*.rmd$", path)])
+  } else {
+    stop("Some file/directory do not exists")
+  }
+
+res <- lapply(all_f, att_from_rmd) %>%
     unlist() %>%
     unique() %>%
     na.omit()
 
-  if (grepl("vignettes", path)) {
+  if (isTRUE(any(grepl("vignettes", path)))) {
     c("knitr", "rmarkdown", res)
   } else {
     res